// src/core/loja/modules/atendimento/utils/log-atividades.enum.ts
export enum TIPO_EVENTO_LOG {
  CRIACAO = 'CRIACAO',
  MUDANCA_STATUS = 'MUDANCA_STATUS',
  ATRIBUICAO_RESPONSAVEL = 'ATRIBUICAO_RESPONSAVEL',
  REMOCAO_RESPONSAVEL = 'REMOCAO_RESPONSAVEL',
  ADICAO_COMENTARIO = 'ADICAO_COMENTARIO',
  AGENDAMENTO_VISITA = 'AGENDAMENTO_VISITA',
  CONFIRMACAO_VISITA = 'CONFIRMACAO_VISITA',
  CONCLUSAO_VISITA = 'CONCLUSAO_VISITA',
  REMOCAO_VISITA = 'REMOCAO_VISITA',
  CRIACAO_TAREFA = 'CRIACAO_TAREFA',
  CONCLUSAO_TAREFA = 'CONCLUSAO_TAREFA',
  REMOCAO_TAREFA = 'REMOCAO_TAREFA',
  ALTERACAO_TIPO_ATENDIMENTO = 'ALTERACAO_TIPO_ATENDIMENTO',
  ALTERACAO_DADOS_CLIENTE = 'ALTERACAO_DADOS_CLIENTE',
  ARQUIVAMENTO = 'ARQUIVAMENTO',
  DESARQUIVAMENTO = 'DESARQUIVAMENTO',
  ALTERACAO_TITULO = 'ALTERACAO_TITULO',
  ALTERACAO_OBSERVACAO = 'ALTERACAO_OBSERVACAO',
  ALTERACAO_TEMPERATURA = 'ALTERACAO_TEMPERATURA',
  ADICAO_ANEXO = 'ADICAO_ANEXO',
  REMOCAO_ANEXO = 'REMOCAO_ANEXO',
  COMPARTILHAMENTO = 'COMPARTILHAMENTO',
  SUSPENSAO_CRIADA = 'SUSPENSAO_CRIADA',
  SUSPENSAO_ATUALIZADA = 'SUSPENSAO_ATUALIZADA',
  SUSPENSAO_REMOVIDA = 'SUSPENSAO_REMOVIDA',
  DISTRIBUICAO_CONFIGURADA = 'DISTRIBUICAO_CONFIGURADA',
  ATENDIMENTOS_REDISTRIBUIDOS = 'ATENDIMENTOS_REDISTRIBUIDOS',
  MONITORAMENTO_EXECUTADO = 'MONITORAMENTO_EXECUTADO',
  ALTERACAO_NOME_TAREFA = 'ALTERACAO_NOME_TAREFA',
  ALTERACAO_OBSERVACOES_TAREFA = 'ALTERACAO_OBSERVACOES_TAREFA',
  ALTERACAO_RESPONSAVEL_TAREFA = 'ALTERACAO_RESPONSAVEL_TAREFA',
  ALTERACAO_HORA_INICIO_TAREFA = 'ALTERACAO_HORA_INICIO_TAREFA',
  ALTERACAO_HORA_FIM_TAREFA = 'ALTERACAO_HORA_FIM_TAREFA',
  ALTERACAO_DATA_TAREFA = 'ALTERACAO_DATA_TAREFA',
}